+++ app/src/app/views/stix/technique/technique-view/technique-view.component.ts
@@ -11,7 +11,7 @@ import { StixViewPage } from '../../stix-view-page';
 })
 export class TechniqueViewComponent extends StixViewPage implements OnInit {
 
-    public editing: boolean = false;
+    // public editing: boolean = false;
 
     public get technique(): Technique { return this.config.object as Technique; }
 
@@ -20,9 +20,9 @@ export class TechniqueViewComponent extends StixViewPage implements OnInit {
     }
 
     ngOnInit() {
-        this.route.queryParams.subscribe(params => {
-            this.editing = params["editing"];
-        });
+        // this.route.queryParams.subscribe(params => {
+        //     this.editing = params["editing"];
+        // });
     }
 
     ngAfterContentChecked() {